Split wallet manager into modules
