+++ rust/src/manager/send_flow_manager.rs
@@ -1,10 +1,13 @@
 pub mod alert_state;
 pub mod amount_or_max;
 pub mod btc_on_change;
+mod coin_control;
 pub mod error;
+mod fee_selection;
 pub mod fiat_on_change;
 mod sanitize;
 pub mod state;
+mod validation;
 
 use std::{sync::Arc, time::Duration};
 
@@ -28,7 +31,7 @@ use alert_state::SendFlowAlertState;
 use amount_or_max::AmountOrMax;
 use backon::{ExponentialBuilder, Retryable};
 use btc_on_change::BtcOnChangeHandler;
-use cove_common::consts::{MIN_SEND_AMOUNT, MIN_SEND_SATS};
+use cove_common::consts::MIN_SEND_SATS;
 use cove_types::{
     WalletId,
     address::AddressWithNetwork,
@@ -36,7 +39,7 @@ use cove_types::{
     fees::{FeeRateOptionWithTotalFee, FeeRateOptions, FeeRateOptionsWithTotalFee, FeeSpeed},
     psbt::Psbt,
     unit::BitcoinUnit,
-    utxo::{Utxo, UtxoList},
+    utxo::Utxo,
 };
 use cove_util::format::NumberFormatter as _;
 use error::SendFlowError;
@@ -67,64 +70,6 @@ const LOCK_STATE_LOAD_FAILED_ERROR_ID: &str = "send_flow_lock_state_load_failed"
 pub enum SetAmountFocusField {
     Amount,
     Address,
-}
-
-fn selected_fee_rate_for_options(
-    fee_options: &FeeRateOptionsWithTotalFee,
-    selected_fee_rate: Option<&Arc<FeeRateOptionWithTotalFee>>,
-) -> Arc<FeeRateOptionWithTotalFee> {
-    let Some(selected_fee_rate) = selected_fee_rate else {
-        return Arc::new(fee_options.medium);
-    };
-
-    match selected_fee_rate.fee_speed {
-        FeeSpeed::Custom { .. } => {
-            fee_options.custom().unwrap_or_else(|| fee_options.medium.into())
-        }
-        FeeSpeed::Fast => fee_options.fast.into(),
-        FeeSpeed::Medium => fee_options.medium.into(),
-        FeeSpeed::Slow => fee_options.slow.into(),
-    }
-}
-
-fn amount_exceeds_spendable_balance(amount: Option<u64>, spendable_balance: Option<u64>) -> bool {
-    let amount = amount.unwrap_or(0);
-    if amount == 0 {
-        return false;
-    }
-
-    let Some(spendable) = spendable_balance else {
-        return false;
-    };
-
-    amount > spendable
-}
-
-fn spendable_balance_for_validation(unlocked_spendable_sats: Option<u64>) -> u64 {
-    // fail closed so amount validation cannot overspend locked or unknown UTXOs
-    unlocked_spendable_sats.unwrap_or(0)
-}
-
-fn unavailable_spendable_balance_alert(
-    unlocked_spendable_sats: Option<u64>,
-    lock_state_load_failed: bool,
-) -> Option<SendFlowAlertState> {
-    if lock_state_load_failed {
-        return Some(SendFlowAlertState::General {
-            title: "Unable to Read Locked Coins".to_string(),
-            message: "Cove could not read the lock state for this wallet. Locked coins are excluded for safety. Please try again shortly.".to_string(),
-        });
-    }
-
-    if unlocked_spendable_sats.is_none() {
-        return Some(SendFlowAlertState::General {
-            title: "Balance Still Loading".to_string(),
-            message: "Cove is still checking which coins are unlocked. Please try again shortly."
-                .to_string(),
-        });
-    }
-
-    None
 }
 
 #[uniffi::export(callback_interface)]
@@ -364,7 +309,10 @@ impl RustSendFlowManager {
     pub fn amount_exceeds_balance(&self) -> bool {
         let state = self.state.lock();
 
-        amount_exceeds_spendable_balance(state.amount_sats, state.unlocked_spendable_sats)
+        validation::amount_exceeds_spendable_balance(
+            state.amount_sats,
+            state.unlocked_spendable_sats,
+        )
     }
 
     #[uniffi::method]
@@ -559,8 +507,8 @@ impl RustSendFlowManager {
         let (spendable_balance, unavailable_balance_alert, is_max_selected) = {
             let state = self.state.lock();
             (
-                spendable_balance_for_validation(state.unlocked_spendable_sats),
-                unavailable_spendable_balance_alert(
+                validation::spendable_balance_for_validation(state.unlocked_spendable_sats),
+                validation::unavailable_spendable_balance_alert(
                     state.unlocked_spendable_sats,
                     state.lock_state_load_failed,
                 ),
@@ -633,24 +581,6 @@ impl RustSendFlowManager {
         }
 
         Some(changed)
-    }
-
-    /// get the custom fee rate option
-    #[uniffi::method]
-    pub async fn get_custom_fee_option(
-        self: &Arc<Self>,
-        fee_rate: Arc<FeeRate>,
-        fee_speed: FeeSpeed,
-    ) -> Result<Arc<FeeRateOptionWithTotalFee>, Error> {
-        let fee_rate = Arc::unwrap_or_clone(fee_rate);
-        let psbt = self.build_psbt(None, None, fee_rate).await?;
-
-        let total_fee = psbt.fee().map_err_str(Error::UnableToGetFeeDetails)?;
-
-        let fee_rate_option =
-            FeeRateOptionWithTotalFee { fee_speed, fee_rate, total_fee: Some(total_fee) };
-
-        Ok(fee_rate_option.into())
     }
 
     // MARK: Action handler
@@ -751,23 +681,6 @@ impl RustSendFlowManager {
 
 // MARK: Private getters
 impl RustSendFlowManager {
-    fn selected_fee_rate(&self) -> Option<Arc<FeeRateOptionWithTotalFee>> {
-        self.state.lock().fee_selection.as_ref().map(|selection| selection.selected.clone())
-    }
-
-    fn fee_rate_options(&self) -> Option<Arc<FeeRateOptionsWithTotalFee>> {
-        self.state.lock().fee_selection.as_ref().map(|selection| selection.options.clone())
-    }
-
-    fn fee_selection_for_options(
-        &self,
-        fee_options: Arc<FeeRateOptionsWithTotalFee>,
-    ) -> FeeSelection {
-        let selected = self.selected_fee_rate();
-        let selected = selected_fee_rate_for_options(&fee_options, selected.as_ref());
-        FeeSelection::new(fee_options, selected)
-    }
-
     pub fn send_amount(&self) -> Option<Amount> {
         let amount_sats = self.state.lock().amount_sats?;
         Some(Amount::from_sat(amount_sats))
@@ -1196,120 +1109,6 @@ impl RustSendFlowManager {
         sender.queue(Message::UpdateFocusField(new));
     }
 
-    fn handle_coin_control_amount_changed(self: &Arc<Self>, amount: f64) -> Option<()> {
-        debug!("handle_coin_control_amount_changed: {amount}");
-
-        let mut coin_control_mode = match self.state.lock().mode.clone() {
-            EnterMode::CoinControl(coin_control_mode) => coin_control_mode,
-            _ => return None,
-        };
-
-        let unit = self.state.lock().metadata.selected_unit;
-        let amount = match unit {
-            BitcoinUnit::Btc => Amount::from_btc(amount).ok()?,
-            BitcoinUnit::Sat => Amount::from_sat(amount as u64),
-        }
-        .max(MIN_SEND_AMOUNT.into());
-
-        // if the amount we are selecting is within 1000 sats of the max send, then select the max send
-        let max_send_without_fees_and_small_utxo = self.max_send_minus_fees_and_small_utxo()?;
-
-        if amount >= max_send_without_fees_and_small_utxo {
-            debug!(
-                "setting coin control to max amount close to max {} {}",
-                amount.as_sats(),
-                max_send_without_fees_and_small_utxo.as_sats()
-            );
-
-            let max_send = coin_control_mode.max_send();
-            coin_control_mode.is_max_selected = true;
-
-            self.state.lock().mode = EnterMode::CoinControl(coin_control_mode);
-            self.handle_amount_changed(max_send);
-            return Some(());
-        }
-
-        // update the state, remove the max selected flag and set the amount
-        {
-            let mut state = self.state.lock();
-            coin_control_mode.is_max_selected = false;
-            state.mode = EnterMode::CoinControl(coin_control_mode);
-        }
-
-        self.handle_amount_changed(amount);
-
-        Some(())
-    }
-
-    fn handle_coin_control_entered_amount_changed(
-        self: &Arc<Self>,
-        amount: String,
-        is_focused: bool,
-    ) -> Option<()> {
-        debug!("handle_coin_control_entered_amount_changed: {amount}");
-        let amount = amount.chars().filter(|c| c.is_numeric() || *c == '.').collect::<String>();
-        let amount_float = amount.parse::<f64>().ok()?;
-
-        if amount_float < MIN_SEND_SATS as f64 && is_focused {
-            return None;
-        }
-
-        self.handle_coin_control_amount_changed(amount_float)
-    }
-
-    fn set_coin_control_mode(self: &Arc<Self>, utxos: Vec<Utxo>) {
-        if utxos.is_empty() {
-            return;
-        }
-
-        match self.state.lock().mode.clone() {
-            // already in coin control mode with the same utxos, so do nothing
-            EnterMode::CoinControl(cc) if cc.utxo_list.utxos == utxos => {
-                return;
-            }
-            _ => {}
-        }
-
-        let utxo_list = Arc::new(UtxoList::from(utxos));
-        let total_minus_fees = {
-            let mut state = self.state.lock();
-            let total_fee_sats = state
-                .fee_selection
-                .as_ref()
-                .and_then(|selection| selection.selected.total_fee.map(|f| f.as_sats()));
-
-            state.mode = EnterMode::coin_control_max(utxo_list.clone());
-            let total_minus_fees =
-                utxo_list.total.as_sats().saturating_sub(total_fee_sats.unwrap_or(1000));
-
-            Amount::from_sat(total_minus_fees)
-        };
-
-        // update the fee rate options for utxos
-        let me = self.clone();
-        cove_tokio::task::spawn(async move {
-            me.get_or_update_fee_rate_options().await;
-        });
-
-        // update the amount
-        self.handle_amount_changed(total_minus_fees);
-    }
-
-    fn disable_coin_control_mode(self: &Arc<Self>) {
-        if !self.state.lock().mode.is_coin_control() {
-            debug!("coin control mode is already disabled");
-            return;
-        }
-
-        self.state.lock().mode = EnterMode::SetAmount;
-
-        // update the fee rate options
-        let me = self.clone();
-        cove_tokio::task::spawn(async move {
-            me.get_or_update_fee_rate_options().await;
-        });
-    }
-
     async fn build_psbt(
         &self,
         address: Option<Address>,
@@ -1481,19 +1280,6 @@ impl RustSendFlowManager {
         }
 
         Ok(())
-    }
-
-    async fn get_and_update_base_fee_rate_options(self: &Arc<Self>) -> Option<Arc<FeeRateOptions>> {
-        let fee_response = FEE_CLIENT.fetch_and_get_fees().await.ok()?;
-        let fees = Arc::new(FeeRateOptions::from(fee_response));
-
-        {
-            let mut state = self.state.lock();
-            state.fee_rate_options_base = Some(fees.clone());
-            state.has_base_fees = true;
-        }
-
-        Some(fees)
     }
 
     fn handle_selected_unit_changed(self: &Arc<Self>, old: BitcoinUnit, new: BitcoinUnit) {
@@ -1877,172 +1663,6 @@ impl RustSendFlowManager {
         });
     }
 
-    async fn get_or_update_fee_rate_options(self: &Arc<Self>) {
-        debug!("get_or_update_fee_rate_options");
-
-        let mut sender = DeferredSender::new(self.reconciler.clone());
-
-        let (address, amount_sats) = {
-            let state = self.state.lock();
-            let address = state.address.clone();
-            let amount_sats = state.amount_sats;
-            (address, amount_sats)
-        };
-
-        debug!("get_or_update_fee_rate_options: {address:?}, {amount_sats:?}");
-        let wallet_actor = self.wallet_actor();
-        let state = self.state.clone();
-
-        let fee_rate_options_base = {
-            let fee_rate_options_base = state.lock().fee_rate_options_base.clone();
-            let fee_rate_options_base = match fee_rate_options_base {
-                Some(fee_rate_options_base) => Some(fee_rate_options_base),
-                None => self.get_and_update_base_fee_rate_options().await,
-            };
-
-            match fee_rate_options_base {
-                Some(fee_rate_options_base) => Arc::unwrap_or_clone(fee_rate_options_base),
-                None => return,
-            }
-        };
-
-        let address = {
-            let first_address = state.lock().first_address.clone();
-            if first_address.is_none() {
-                let () = self.get_first_address().await;
-            }
-
-            match (address, first_address) {
-                (Some(address), _) => address,
-                (None, Some(first_address)) => first_address,
-                _ => return,
-            }
-        };
-
-        let mode = self.state.lock().mode.clone();
-        let address = Arc::unwrap_or_clone(address);
-
-        let amount_sats_for_fee_calc = match &mode {
-            EnterMode::CoinControl(cc) if cc.is_max_selected => cc.max_send().to_sat(),
-            _ => amount_sats.unwrap_or(MIN_SEND_SATS),
-        };
-
-        let amount_for_fee_calc = Amount::from_sat(amount_sats_for_fee_calc);
-        let max_selected = self.state.lock().max_selected.clone();
-
-        let new_fee_rate_options = match (max_selected, &mode) {
-            (Some(_), _) => {
-                call!(wallet_actor.fee_rate_options_with_total_fee_for_drain(
-                    fee_rate_options_base,
-                    address.clone()
-                ))
-            }
-            (None, EnterMode::SetAmount) => {
-                call!(wallet_actor.fee_rate_options_with_total_fee(
-                    fee_rate_options_base,
-                    amount_for_fee_calc.into(),
-                    address.clone()
-                ))
-            }
-            (None, EnterMode::CoinControl(cc)) => {
-                call!(wallet_actor.fee_rate_options_with_total_fee_for_manual(
-                    cc.utxo_list(),
-                    fee_rate_options_base,
-                    amount_for_fee_calc.into(),
-                    address.clone()
-                ))
-            }
-        }
-        .await
-        .unwrap();
-
-        let mut fee_rate_options = match new_fee_rate_options {
-            Ok(fee_rate_options) => fee_rate_options,
-            Err(_) => return,
-        };
-
-        // if user had a custom speed selected, re-apply it
-        let selected_fee_rate = self.selected_fee_rate();
-        if let Some(updated_options) = self
-            .updated_custom_fee_option(
-                address.clone(),
-                amount_for_fee_calc,
-                fee_rate_options,
-                selected_fee_rate.clone(),
-            )
-            .await
-        {
-            fee_rate_options = updated_options;
-        }
-
-        // update the state
-        let fee_rate_options_with_total_fee = Arc::new(fee_rate_options);
-        let selected = selected_fee_rate_for_options(
-            &fee_rate_options_with_total_fee,
-            selected_fee_rate.as_ref(),
-        );
-        let fee_selection = FeeSelection::new(fee_rate_options_with_total_fee, selected);
-        state.lock().fee_selection = Some(fee_selection.clone());
-
-        // update the amount if in coin control max send
-        match &mode {
-            EnterMode::CoinControl(cc) if cc.is_max_selected => {
-                let max = cc.max_send();
-                let total_fee = fee_selection
-                    .selected
-                    .total_fee
-                    .map(|fee| fee.as_sats())
-                    .or_else(|| fee_selection.options.medium.total_fee.map(|fee| fee.as_sats()))
-                    .unwrap_or(0);
-
-                let send_amount = max.as_sats() - total_fee;
-                if Some(send_amount) != amount_sats {
-                    self.handle_amount_changed(Amount::from_sat(send_amount));
-                }
-            }
-            _ => {}
-        }
-
-        sender.queue(Message::UpdateFeeSelection(fee_selection));
-    }
-
-    /// Returns the fee rate options with the updated custom fee
-    async fn updated_custom_fee_option(
-        self: &Arc<Self>,
-        address: Address,
-        amount: Amount,
-        fee_rate_options: FeeRateOptionsWithTotalFee,
-        selected_fee_rate: Option<Arc<FeeRateOptionWithTotalFee>>,
-    ) -> Option<FeeRateOptionsWithTotalFee> {
-        // only update if the selected fee rate is custom
-        let selected_fee_rate = selected_fee_rate?;
-        if !matches!(selected_fee_rate.fee_speed, FeeSpeed::Custom { .. }) {
-            return None;
-        }
-
-        let psbt = self
-            .build_psbt(Some(address), Some(amount), selected_fee_rate.fee_rate)
-            .await
-            .map_err_str(Error::UnableToGetFeeDetails);
-
-        let total_fee = psbt.and_then(|psbt| psbt.fee().map_err_str(Error::UnableToGetFeeDetails));
-
-        let total_fee = match total_fee {
-            Ok(total_fee) => total_fee,
-            Err(error) => {
-                let error = SendFlowError::UnableToGetMaxSend(error.to_string());
-                self.reconciler.send_async(Message::SetAlert(error.into())).await;
-                return None;
-            }
-        };
-
-        let mut new_custom_with_fee = Arc::unwrap_or_clone(selected_fee_rate.clone());
-        new_custom_with_fee.total_fee = Some(total_fee);
-
-        let fee_rate_options = fee_rate_options.add_custom_fee_rate(new_custom_with_fee.into());
-        Some(fee_rate_options)
-    }
-
     async fn get_wallet_balance(self: &Arc<Self>) {
         let balance = self.wallet_manager.balance().await;
         let unlocked_spendable_sats =
@@ -2098,42 +1718,6 @@ mod tests {
                 super::Duration::from_millis(200),
             ),
         })
-    }
-
-    #[test]
-    fn amount_exceeds_spendable_balance_uses_unlocked_balance() {
-        assert!(super::amount_exceeds_spendable_balance(Some(6_000), Some(5_000)));
-        assert!(!super::amount_exceeds_spendable_balance(Some(5_000), Some(5_000)));
-        assert!(!super::amount_exceeds_spendable_balance(Some(0), Some(0)));
-        assert!(!super::amount_exceeds_spendable_balance(Some(1), None));
-        assert!(!super::amount_exceeds_spendable_balance(Some(0), None));
-    }
-
-    #[test]
-    fn validation_spendable_balance_uses_zero_when_unlocked_balance_is_unknown() {
-        assert_eq!(super::spendable_balance_for_validation(Some(5_000)), 5_000);
-        assert_eq!(super::spendable_balance_for_validation(None), 0);
-    }
-
-    #[test]
-    fn unavailable_balance_alert_distinguishes_lock_state_failures() {
-        let alert = super::unavailable_spendable_balance_alert(None, true);
-
-        assert!(matches!(
-            alert,
-            Some(super::SendFlowAlertState::General { title, .. }) if title.contains("Locked")
-        ));
-    }
-
-    #[test]
-    fn unavailable_balance_alert_distinguishes_loading_state() {
-        let alert = super::unavailable_spendable_balance_alert(None, false);
-
-        assert!(matches!(
-            alert,
-            Some(super::SendFlowAlertState::General { title, .. }) if title.contains("Loading")
-        ));
-        assert_eq!(super::unavailable_spendable_balance_alert(Some(5_000), false), None);
     }
 
     #[test]
