+++ rust/src/manager/wallet_manager/actor.rs
@@ -1,66 +1,34 @@
 use crate::{
-    database::{
-        Database,
-        wallet_data::{ReceiveAddressCache, WalletDataDb},
-    },
+    database::{Database, wallet_data::WalletDataDb},
     historical_price_service::HistoricalPriceService,
     manager::wallet_manager::{
-        Error, SendFlowErrorAlert, TransactionLockState, WalletLedgerState, WalletManagerError,
-        WalletScanPhase, WalletScanStatus,
-        receive_address::{
-            CACHE_WINDOW, ReceiveAddressPresentation, ReceiveAddressRefreshState,
-            ReceiveAddressSession, ReceiveAddressState, ReceiveAddressStatus,
-            RefreshExpiredAddressDecision,
-        },
+        Error, SendFlowErrorAlert, TransactionLockState, WalletLedgerState, WalletScanPhase,
+        WalletScanStatus, receive_address::ReceiveAddressSession,
     },
-    mnemonic,
-    node::{
-        Node,
-        client::{Error as NodeError, NodeClient, NodeClientOptions},
-        client_builder::NodeClientBuilder,
-    },
+    node::client::{Error as NodeError, NodeClient},
     receive_address_watcher::ReceiveAddressWatcher,
-    transaction::{ConfirmedTransaction, FeeRate, Transaction, TransactionDetails, TxId},
+    transaction::{ConfirmedTransaction, Transaction, TransactionDetails, TxId},
     transaction_watcher::TransactionWatcher,
-    wallet::{
-        Address, AddressInfo, Wallet, WalletAddressType, balance::Balance, metadata::BlockSizeLast,
-        metadata::WalletMetadata,
-    },
+    wallet::{Wallet, WalletAddressType, balance::Balance, metadata::WalletMetadata},
 };
+mod node;
+mod receive_address;
 mod scan;
+mod transaction_confirmation;
+mod transactions;
 
-use super::payjoin::{PayjoinActor, PayjoinSender, build_sender};
+use super::payjoin::PayjoinActor;
 use act_zero::{runtimes::tokio::spawn_actor, *};
 use act_zero_ext::into_actor_result;
 use ahash::HashMap;
 use bdk_wallet::{
-    AddUtxoError, Utxo, WeightedUtxo,
-    chain::{
-        BlockId, ChainPosition, TxGraph,
-        bitcoin::Psbt,
-        spk_client::{FullScanResponse, SyncRequest, SyncResponse},
-    },
-    error::CreateTxError,
+    KeychainKind, LocalOutput,
+    chain::{ChainPosition, spk_client::FullScanResponse},
     tx_builder::TxBuilder,
 };
-// Note: SignOptions is marked deprecated in bdk_wallet 2.2.0 with a misleading message
-// saying it moved to bitcoin::psbt, but no replacement exists there yet. bdk_wallet itself
-// uses #![allow(deprecated)] and still requires SignOptions in its API. This will be resolved
-// when bdk_wallet completes their signer refactoring.
-#[allow(deprecated)]
-use bdk_wallet::SignOptions;
-use bdk_wallet::{KeychainKind, LocalOutput, TxOrdering};
-use bitcoin::{Amount, FeeRate as BdkFeeRate, OutPoint, TxIn, Txid, constants::COINBASE_MATURITY};
-use bitcoin::{Transaction as BdkTransaction, params::Params};
-use cove_bdk::coin_selection::CoveDefaultCoinSelection;
+use bitcoin::{Amount, OutPoint, Txid, constants::COINBASE_MATURITY};
 use cove_bdk_progressive_scan::ScanUpdate;
-use cove_common::consts::MIN_SEND_AMOUNT;
-use cove_tokio::{AbortableTask, FutureTimeoutExt as _};
-use cove_types::{
-    confirm::{AddressAndAmount, ConfirmDetails, ExtraItem, InputOutputDetails, SplitOutput},
-    fees::{FeeRateOptionWithTotalFee, FeeRateOptions, FeeRateOptionsWithTotalFee},
-    utxo::{UtxoList, UtxoType},
-};
+use cove_tokio::AbortableTask;
 use cove_util::result_ext::ResultExt as _;
 use eyre::Result;
 use flume::Sender;
@@ -71,10 +39,7 @@ use std::{
     sync::Arc,
     time::{Duration, UNIX_EPOCH},
 };
-use tap::TapFallible as _;
-use tracing::{debug, error, info, warn};
-
-use self::mnemonic::{Mnemonic, MnemonicExt as _};
+use tracing::{debug, error, warn};
 
 use self::scan::{
     EMPTY_WALLET_SCAN_PROGRESS_DELAY, FullScanType, PreparedProgressiveScan,
@@ -82,8 +47,6 @@ use self::scan::{
     WalletScanEvent, WalletScanEventKind, should_update_full_scan_metadata,
 };
 use super::{SingleOrMany, WalletManagerReconcileMessage};
-
-const RECEIVE_ADDRESS_FRESHNESS_TIMEOUT: Duration = Duration::from_millis(400);
 
 #[derive(Debug)]
 pub struct WalletActor {
@@ -222,128 +185,6 @@ impl WalletActor {
         self.unlocked_trusted_spendable_balance_inner()
     }
 
-    // Build a transaction but don't advance the change address index
-    #[into_actor_result]
-    pub async fn build_ephemeral_drain_tx(
-        &mut self,
-        address: Address,
-        fee: FeeRate,
-    ) -> Result<Psbt, Error> {
-        self.ensure_ledger_ready_for_spend()?;
-
-        debug!("build_ephemeral_drain_tx for fee rate {}", fee.sat_per_vb());
-        let script_pubkey = address.script_pubkey();
-        let locked_outpoints = self.locked_output_outpoints()?;
-        let mut tx_builder = self.wallet.bdk.build_tx();
-
-        exclude_locked_outpoints(&mut tx_builder, locked_outpoints);
-        tx_builder.drain_wallet().drain_to(script_pubkey).fee_rate(fee.into());
-        let psbt = tx_builder.finish().map_err_str(Error::BuildTxError)?;
-        self.wallet.unreserve_tx_change_addresses(&psbt.unsigned_tx);
-
-        Ok(psbt)
-    }
-
-    /// Build a transaction
-    #[into_actor_result]
-    pub async fn build_tx(
-        &mut self,
-        amount: Amount,
-        address: Address,
-        fee_rate: impl Into<BdkFeeRate>,
-    ) -> Result<Psbt, Error> {
-        self.ensure_ledger_ready_for_spend()?;
-
-        debug!("build_tx");
-        let fee_rate = fee_rate.into();
-        let script_pubkey = address.script_pubkey();
-
-        let coin_selection = CoveDefaultCoinSelection::new(self.seed);
-        let locked_outpoints = self.locked_output_outpoints()?;
-        let mut tx_builder = self.wallet.bdk.build_tx().coin_selection(coin_selection);
-
-        exclude_locked_outpoints(&mut tx_builder, locked_outpoints);
-        tx_builder.ordering(TxOrdering::Untouched);
-        tx_builder.add_recipient(script_pubkey, amount);
-        tx_builder.fee_rate(fee_rate);
-
-        let psbt = tx_builder.finish().map_err_str(Error::BuildTxError)?;
-        Ok(psbt)
-    }
-
-    /// Build a transaction but don't advance the change address index
-    #[into_actor_result]
-    pub async fn build_ephemeral_tx(
-        &mut self,
-        amount: Amount,
-        address: Address,
-        fee: impl Into<BdkFeeRate>,
-    ) -> Result<Psbt, Error> {
-        self.ensure_ledger_ready_for_spend()?;
-
-        debug!("build_ephemeral_tx");
-        let psbt = self.do_build_tx(amount, address, fee).await?;
-        self.wallet.unreserve_tx_change_addresses(&psbt.unsigned_tx);
-        Ok(psbt)
-    }
-
-    /// Build a transaction using only the given UTXOs
-    #[into_actor_result]
-    pub async fn build_manual_tx(
-        &mut self,
-        utxos: Vec<OutPoint>,
-        total_amount: Amount,
-        address: Address,
-        fee_rate: impl Into<BdkFeeRate>,
-    ) -> Result<Psbt, Error> {
-        self.ensure_ledger_ready_for_spend()?;
-
-        debug!("build_manual_tx: {total_amount:?}");
-
-        let fee_rate = fee_rate.into();
-        let send_amount = self.get_max_send_for_utxos(total_amount, &address, fee_rate, &utxos)?;
-
-        let mut tx_builder = self.wallet.bdk.build_tx();
-        tx_builder.add_utxos(&utxos).map_err_str(Error::AddUtxosError)?;
-        tx_builder.manually_selected_only();
-        tx_builder.ordering(TxOrdering::Untouched);
-        tx_builder.add_recipient(address.script_pubkey(), send_amount);
-        tx_builder.fee_rate(fee_rate);
-
-        let psbt = tx_builder.finish().map_err_str(Error::BuildTxError)?;
-        Ok(psbt)
-    }
-
-    /// Build a manual transaction but don't advance the change address index
-    #[into_actor_result]
-    pub async fn build_manual_ephemeral_tx(
-        &mut self,
-        utxos: Vec<OutPoint>,
-        amount: Amount,
-        address: Address,
-        fee: impl Into<BdkFeeRate>,
-    ) -> Result<Psbt, Error> {
-        self.ensure_ledger_ready_for_spend()?;
-
-        debug!("build_manual_ephemeral_tx");
-        let psbt = self.do_build_manual_tx(utxos, amount, address, fee).await?;
-        self.wallet.unreserve_tx_change_addresses(&psbt.unsigned_tx);
-        Ok(psbt)
-    }
-
-    // release reserved change addresses for a discarded unsigned transaction
-    pub async fn cancel_txn(&mut self, txn: BdkTransaction) {
-        self.wallet.unreserve_tx_change_addresses(&txn);
-    }
-
-    pub async fn list_unspent(&mut self) -> ActorResult<Result<Vec<LocalOutput>, Error>> {
-        if let Err(error) = self.ensure_ledger_ready_for_spend() {
-            return Produces::ok(Err(error));
-        }
-
-        Produces::ok(Ok(self.wallet.bdk.list_unspent().collect()))
-    }
-
     #[act_zero_ext::into_actor_result]
     pub async fn transactions(&mut self) -> Vec<Transaction> {
         let zero = Amount::ZERO.into();
@@ -367,727 +208,6 @@ impl WalletActor {
 
         transactions.sort_unstable_by(|a, b| a.cmp(b).reverse());
         transactions
-    }
-
-    pub async fn split_transaction_outputs(
-        &mut self,
-        outputs: Vec<AddressAndAmount>,
-    ) -> ActorResult<SplitOutput> {
-        let external = outputs
-            .iter()
-            .filter(|output| !self.wallet.bdk.is_mine(output.address.script_pubkey()))
-            .cloned()
-            .collect();
-
-        let internal = outputs
-            .iter()
-            .filter(|output| self.wallet.bdk.is_mine(output.address.script_pubkey()))
-            .cloned()
-            .collect();
-
-        Produces::ok(SplitOutput { external, internal })
-    }
-
-    #[into_actor_result]
-    pub async fn fee_rate_options_with_total_fee(
-        &mut self,
-        fee_rate_options: FeeRateOptions,
-        amount: Amount,
-        address: Address,
-    ) -> Result<FeeRateOptionsWithTotalFee, Error> {
-        self.ensure_ledger_ready_for_spend()?;
-
-        let fast_fee_rate = fee_rate_options.fast.fee_rate;
-        let medium_fee_rate = fee_rate_options.medium.fee_rate;
-        let slow_fee_rate = fee_rate_options.slow.fee_rate;
-
-        let fast_psbt = self.do_build_ephemeral_tx(amount, address.clone(), fast_fee_rate).await?;
-
-        let medium_psbt =
-            self.do_build_ephemeral_tx(amount, address.clone(), medium_fee_rate).await?;
-
-        let slow_psbt = self.do_build_ephemeral_tx(amount, address.clone(), slow_fee_rate).await?;
-
-        let options = FeeRateOptionsWithTotalFee {
-            fast: FeeRateOptionWithTotalFee::new(
-                fee_rate_options.fast,
-                fast_psbt.fee().map_err_str(Error::FeesError)?,
-            ),
-            medium: FeeRateOptionWithTotalFee::new(
-                fee_rate_options.medium,
-                medium_psbt.fee().map_err_str(Error::FeesError)?,
-            ),
-            slow: FeeRateOptionWithTotalFee::new(
-                fee_rate_options.slow,
-                slow_psbt.fee().map_err_str(Error::FeesError)?,
-            ),
-            custom: None,
-        };
-
-        Ok(options)
-    }
-
-    #[into_actor_result]
-    pub async fn fee_rate_options_with_total_fee_for_manual(
-        &mut self,
-        utxos: Arc<UtxoList>,
-        fee_rate_options: FeeRateOptions,
-        amount: Amount,
-        address: Address,
-    ) -> Result<FeeRateOptionsWithTotalFee, Error> {
-        self.ensure_ledger_ready_for_spend()?;
-
-        debug!("fee_rate_options_with_total_fee_for_manual");
-        let fast_fee_rate = fee_rate_options.fast.fee_rate;
-        let medium_fee_rate = fee_rate_options.medium.fee_rate;
-        let slow_fee_rate = fee_rate_options.slow.fee_rate;
-
-        let fast_psbt = self
-            .do_build_manual_ephemeral_tx(
-                utxos.clone().outpoints(),
-                amount,
-                address.clone(),
-                fast_fee_rate,
-            )
-            .await?;
-
-        let medium_psbt = self
-            .do_build_manual_ephemeral_tx(
-                utxos.clone().outpoints(),
-                amount,
-                address.clone(),
-                medium_fee_rate,
-            )
-            .await?;
-
-        let slow_psbt = self
-            .do_build_manual_ephemeral_tx(
-                utxos.clone().outpoints(),
-                amount,
-                address.clone(),
-                slow_fee_rate,
-            )
-            .await?;
-
-        let options = FeeRateOptionsWithTotalFee {
-            fast: FeeRateOptionWithTotalFee::new(
-                fee_rate_options.fast,
-                fast_psbt.fee().map_err_str(Error::FeesError)?,
-            ),
-            medium: FeeRateOptionWithTotalFee::new(
-                fee_rate_options.medium,
-                medium_psbt.fee().map_err_str(Error::FeesError)?,
-            ),
-            slow: FeeRateOptionWithTotalFee::new(
-                fee_rate_options.slow,
-                slow_psbt.fee().map_err_str(Error::FeesError)?,
-            ),
-            custom: None,
-        };
-
-        Ok(options)
-    }
-
-    #[into_actor_result]
-    pub async fn fee_rate_options_with_total_fee_for_drain(
-        &mut self,
-        fee_rate_options: FeeRateOptions,
-        address: Address,
-    ) -> Result<FeeRateOptionsWithTotalFee, Error> {
-        self.ensure_ledger_ready_for_spend()?;
-
-        let options = fee_rate_options;
-
-        let fast_fee_rate = options.fast.fee_rate;
-        let fast_psbt: Psbt =
-            self.do_build_ephemeral_drain_tx(address.clone(), fast_fee_rate).await?;
-
-        let medium_fee_rate = options.medium.fee_rate;
-        let medium_psbt: Psbt =
-            self.do_build_ephemeral_drain_tx(address.clone(), medium_fee_rate).await?;
-
-        let slow_fee_rate = options.slow.fee_rate;
-        let slow_psbt: Psbt =
-            self.do_build_ephemeral_drain_tx(address.clone(), slow_fee_rate).await?;
-
-        let options_with_fee = FeeRateOptionsWithTotalFee {
-            fast: FeeRateOptionWithTotalFee::new(
-                options.fast,
-                fast_psbt.fee().map_err_str(Error::FeesError)?,
-            ),
-            medium: FeeRateOptionWithTotalFee::new(
-                options.medium,
-                medium_psbt.fee().map_err_str(Error::FeesError)?,
-            ),
-            slow: FeeRateOptionWithTotalFee::new(
-                options.slow,
-                slow_psbt.fee().map_err_str(Error::FeesError)?,
-            ),
-            custom: None,
-        };
-
-        Ok(options_with_fee)
-    }
-
-    #[into_actor_result]
-    pub async fn get_confirm_details(
-        &mut self,
-        psbt: Psbt,
-        fee_rate: BdkFeeRate,
-    ) -> Result<ConfirmDetails, Error> {
-        self.ensure_ledger_ready_for_spend()?;
-
-        #[inline(always)]
-        fn error(s: &str) -> WalletManagerError {
-            WalletManagerError::GetConfirmDetailsError(s.to_string())
-        }
-
-        #[inline(always)]
-        fn err_fmt(s: &str, err: impl std::fmt::Display) -> WalletManagerError {
-            WalletManagerError::GetConfirmDetailsError(format!("{s}: {err}"))
-        }
-
-        let network = self.wallet.bdk.network();
-        let external_outputs = psbt
-            .unsigned_tx
-            .output
-            .iter()
-            .filter(|output| !self.wallet.bdk.is_mine(output.script_pubkey.clone()))
-            .collect::<Vec<&bitcoin::TxOut>>();
-
-        if external_outputs.len() > 1 {
-            return Err(error("multiple address to send to found, not currently supported"));
-        }
-
-        // if there is an external output, use that
-        // otherwise this is a consolidation txn, sending to the same wallet so use the first output
-        let output = external_outputs
-            .first()
-            .copied()
-            .or_else(|| psbt.unsigned_tx.output.first())
-            .ok_or_else(|| error("no addess to send to found"))?;
-
-        let params = Params::from(network);
-        let sending_to = bitcoin::Address::from_script(&output.script_pubkey, params)
-            .map_err(|err| err_fmt("unable to get address from script", err))?;
-
-        let output_amount = output.value;
-
-        if output_amount == Amount::ZERO {
-            return Err(error("zero amount"));
-        }
-
-        let fee = psbt.fee().map_err(|err| err_fmt("unable to get fee", err))?;
-
-        let total_spent = output_amount
-            .checked_add(fee)
-            .ok_or_else(|| error("fee overflow, cannot calculate spending amount"))?;
-
-        let psbt = cove_types::psbt::Psbt::from(psbt);
-        let labels_db = self.db.labels.clone();
-
-        let extras = psbt
-            .utxos_iter()
-            .map(|(tx_in, tx_out)| {
-                let outpoint = &tx_in.previous_output;
-                let utxo_type = self.wallet.bdk.get_utxo(*outpoint).map(|x| match x.keychain {
-                    KeychainKind::External => UtxoType::Output,
-                    KeychainKind::Internal => UtxoType::Change,
-                });
-
-                let address =
-                    bitcoin::Address::from_script(&tx_out.script_pubkey, Params::from(network))
-                        .ok();
-
-                let label = labels_db
-                    .get_txn_label_record(outpoint.txid)
-                    .ok()
-                    .flatten()
-                    .map(|record| record.item.label)
-                    .unwrap_or_else(|| match address {
-                        Some(address) => labels_db
-                            .get_address_record(address.into_unchecked())
-                            .ok()
-                            .flatten()
-                            .and_then(|record| record.item.label),
-                        None => None,
-                    });
-
-                let extra = ExtraItem::new(label, utxo_type);
-                (&outpoint.txid, extra)
-            })
-            .collect();
-
-        let more_details = InputOutputDetails::new_with_labels(&psbt, self.wallet.network, extras);
-        let fee_percentage = fee.to_sat() * 100 / output_amount.to_sat();
-
-        let details = ConfirmDetails {
-            spending_amount: total_spent.into(),
-            sending_amount: output_amount.into(),
-            fee_total: fee.into(),
-            fee_rate: fee_rate.into(),
-            fee_percentage,
-            sending_to: sending_to.into(),
-            psbt,
-            more_details,
-        };
-
-        Ok(details)
-    }
-
-    pub async fn initiate_payment(
-        &mut self,
-        psbt: Psbt,
-        payjoin_endpoint: Option<String>,
-    ) -> ActorResult<Result<(), Error>> {
-        if let Err(error) = self.ensure_ledger_ready_for_spend() {
-            return Produces::ok(Err(error));
-        }
-
-        if self.payjoin_actor.is_some() {
-            return Produces::ok(Err(Error::SignAndBroadcastError(
-                "a payjoin session is already in progress".to_string(),
-            )));
-        }
-
-        match payjoin_endpoint {
-            // non-payjoin: sign and broadcast the original transaction directly
-            None => {
-                let result = self.do_sign_and_broadcast_transaction(psbt).await;
-                Produces::ok(result)
-            }
-            // payjoin: initiate BIP77 v2 OHTTP negotiation with the receiver
-            Some(endpoint) => {
-                let result = self.initiate_payjoin_payment(psbt, endpoint).await;
-                Produces::ok(result)
-            }
-        }
-    }
-
-    async fn initiate_payjoin_payment(
-        &mut self,
-        psbt: Psbt,
-        endpoint: String,
-    ) -> Result<(), Error> {
-        // sign the original PSBT first so we have both the
-        // signed PSBT to POST to the directory and a valid fallback tx
-        let (signed_psbt, fallback_tx) = self.do_sign_original_psbt(psbt).await?;
-
-        let network: bitcoin::Network = self.wallet.network.into();
-
-        let Ok(sender) = build_sender(signed_psbt, &fallback_tx, endpoint, network)
-            .inspect_err(|e| warn!("payjoin setup failed, broadcasting fallback tx: {e}"))
-        else {
-            send!(self.addr.handle_payjoin_fallback(fallback_tx));
-            return Ok(());
-        };
-
-        self.spawn_payjoin_actor(sender, fallback_tx);
-        Ok(())
-    }
-
-    /// Sign a PSBT with the wallet's mnemonic, returning the signed PSBT and the extracted fallback tx.
-    async fn do_sign_original_psbt(
-        &mut self,
-        mut psbt: Psbt,
-    ) -> Result<(Psbt, BdkTransaction), Error> {
-        fn err(s: &str) -> Error {
-            Error::SignAndBroadcastError(s.to_string())
-        }
-
-        let network = self.wallet.network;
-        let mnemonic = Mnemonic::try_from_id(&self.wallet.metadata.id)
-            .tap_err(|error| error!("failed to get mnemonic for wallet: {error}"))
-            .map_err(|_| err("failed to get mnemonic for wallet"))?;
-
-        let descriptors =
-            mnemonic.into_descriptors(None, network, self.wallet.metadata.address_type);
-
-        let create_params = descriptors.into_create_params().network(network.into());
-
-        // create a new temp wallet with the descriptors to sign
-        let wallet = create_params
-            .create_wallet_no_persist()
-            .tap_err(|error| error!("failed to create wallet: {error}"))
-            .map_err(|_| err("unable to sign"))?;
-
-        let finalized = wallet
-            .sign(&mut psbt, Default::default())
-            .tap_err(|error| error!("failed to sign: {error}"))
-            .map_err(|_| err("unable to sign"))?;
-
-        if !finalized {
-            return Err(err("transaction not finalized, unable to sign"));
-        }
-
-        // clone before extract so we retain the signed PSBT for the payjoin POST
-        let fallback_tx = psbt
-            .clone()
-            .extract_tx_fee_rate_limit()
-            .tap_err(|error| error!("failed to extract transaction: {error}"))
-            .map_err(|_| err("failed to extract transaction"))?;
-
-        Ok((psbt, fallback_tx))
-    }
-
-    async fn do_sign_and_broadcast_transaction(&mut self, psbt: Psbt) -> Result<(), Error> {
-        let (_, transaction) = self.do_sign_original_psbt(psbt).await?;
-
-        self.do_broadcast_transaction(transaction.clone()).await?;
-
-        // insert into local wallet and update UI immediately
-        self.insert_broadcast_transaction(transaction).await;
-
-        Ok(())
-    }
-
-    #[into_actor_result]
-    pub async fn broadcast_transaction(
-        &mut self,
-        transaction: BdkTransaction,
-    ) -> Result<(), Error> {
-        self.ensure_ledger_ready_for_spend()?;
-
-        self.check_node_connection().await.map_err(|error| {
-            let error_string =
-                format!("failed to broadcast transaction, unable to connect to node: {error:?}");
-            Error::SignAndBroadcastError(error_string)
-        })?;
-
-        self.node_client()
-            .await
-            .map_err(|_| {
-                Error::SignAndBroadcastError(
-                    "failed to broadcast transaction, could not get node client, try again"
-                        .to_string(),
-                )
-            })?
-            .broadcast_transaction(transaction.clone())
-            .await
-            .map_err(|error| {
-                let error_string = format!("failed to broadcast transaction, try again: {error:?}");
-                Error::SignAndBroadcastError(error_string)
-            })?;
-
-        // insert into local wallet and update UI immediately
-        self.insert_broadcast_transaction(transaction).await;
-
-        Ok(())
-    }
-
-    #[into_actor_result]
-    #[allow(deprecated)] // SignOptions usage required by bdk_wallet API, no replacement yet
-    pub async fn finalize_psbt(&mut self, psbt: Psbt) -> Result<bitcoin::Transaction, Error> {
-        self.ensure_ledger_ready_for_spend()?;
-
-        let mut psbt = psbt;
-
-        let finalized = self
-            .wallet
-            .bdk
-            .finalize_psbt(&mut psbt, SignOptions::default())
-            .map_err_str(Error::PsbtFinalizeError)?;
-
-        if !finalized {
-            return Err(Error::PsbtFinalizeError("Failed to finalize PSBT".to_string()));
-        }
-
-        let tx = psbt.extract_tx().map_err(|e| {
-            let err = format!("Failed to extract tx from PSBT: {e:?}");
-            Error::PsbtFinalizeError(err)
-        })?;
-
-        Ok(tx)
-    }
-
-    /// Insert a broadcast transaction into the local wallet and update the UI
-    async fn insert_broadcast_transaction(&mut self, transaction: BdkTransaction) {
-        use WalletManagerReconcileMessage as Msg;
-        use std::time::SystemTime;
-
-        let now =
-            SystemTime::now().duration_since(UNIX_EPOCH).map(|d| d.as_secs()).unwrap_or_else(|e| {
-                warn!("System clock skew detected: {e}");
-                u64::MAX
-            });
-        let txid = transaction.compute_txid();
-
-        // insert the unconfirmed transaction into the local wallet
-        self.wallet.bdk.apply_unconfirmed_txs([(transaction, now)]);
-
-        // persist the wallet to save the new transaction
-        if let Err(error) = self.wallet.persist() {
-            error!("Failed to persist wallet after inserting broadcast tx: {error}");
-        }
-
-        // send updated balance to UI
-        let balance = self.wallet.balance();
-        self.send(Msg::WalletBalanceChanged(balance.into()));
-
-        // send updated transactions to UI
-        match self.transactions().await {
-            Ok(future) => match future.await {
-                Ok(transactions) => self.send(Msg::UpdatedTransactions(transactions)),
-                Err(e) => error!("Failed to get transactions after broadcast: {e}"),
-            },
-            Err(e) => error!("Failed to get transactions after broadcast: {e}"),
-        }
-
-        // start a transaction watcher to track confirmations
-        send!(self.addr.start_transaction_watcher(txid));
-    }
-
-    /// Broadcasts `tx`, inserts it into the local wallet, and sends the appropriate reconcile message.
-    async fn broadcast_and_notify(&mut self, tx: BdkTransaction) {
-        match self.do_broadcast_transaction(tx.clone()).await {
-            Ok(()) => {
-                self.insert_broadcast_transaction(tx).await;
-                self.send(WalletManagerReconcileMessage::PayjoinTxBroadcast);
-            }
-            Err(error) => {
-                error!("payjoin broadcast failed: {error}");
-                self.send(WalletManagerReconcileMessage::SendFlowError(
-                    SendFlowErrorAlert::SignAndBroadcast(error.to_string()),
-                ));
-            }
-        }
-    }
-
-    /// Called when the payjoin receiver accepted our proposal and returned a signed PSBT.
-    /// Signs the proposal with our keys, broadcasts it, and notifies the platform.
-    /// Falls back to broadcasting the original transaction if proposal signing or broadcast fails.
-    pub async fn handle_payjoin_success(
-        &mut self,
-        proposal_psbt: Psbt,
-        fallback_tx: BdkTransaction,
-    ) -> ActorResult<()> {
-        let Ok((_, proposal_tx)) =
-            self.do_sign_original_psbt(proposal_psbt).await.inspect_err(|e| {
-                error!("failed to sign payjoin proposal, falling back to original tx: {e:?}")
-            })
-        else {
-            self.broadcast_and_notify(fallback_tx).await;
-            self.payjoin_actor = None;
-            return Produces::ok(());
-        };
-
-        match self.do_broadcast_transaction(proposal_tx.clone()).await {
-            Ok(()) => {
-                self.insert_broadcast_transaction(proposal_tx).await;
-                self.send(WalletManagerReconcileMessage::PayjoinTxBroadcast);
-            }
-            Err(error) => {
-                error!(
-                    "failed to broadcast payjoin proposal, falling back to original tx: {error:?}"
-                );
-                self.broadcast_and_notify(fallback_tx).await;
-            }
-        }
-
-        self.payjoin_actor = None;
-        Produces::ok(())
-    }
-
-    /// Called when payjoin negotiation failed or timed out.
-    /// Broadcasts the original signed transaction as a fallback and notifies the platform.
-    pub async fn handle_payjoin_fallback(
-        &mut self,
-        fallback_tx: BdkTransaction,
-    ) -> ActorResult<()> {
-        self.broadcast_and_notify(fallback_tx).await;
-        self.payjoin_actor = None;
-        Produces::ok(())
-    }
-
-    pub async fn address_at(&mut self, index: u32) -> ActorResult<AddressInfo> {
-        let address = self.wallet.bdk.peek_address(KeychainKind::External, index);
-        Produces::ok(address.into())
-    }
-
-    pub async fn open_receive_address_intent(&mut self) -> ActorResult<()> {
-        self.set_receive_address_loading(true);
-
-        match self.do_open_receive_address().await {
-            Ok(_) => self.set_receive_address_loading(false),
-            Err(error) => {
-                self.set_receive_address_loading(false);
-                self.send(WalletManagerReconcileMessage::ReceiveAddressError(error.to_string()));
-            }
-        }
-
-        Produces::ok(())
-    }
-
-    async fn do_open_receive_address(&mut self) -> Result<ReceiveAddressState, Error> {
-        let now = current_epoch_secs();
-
-        let Some(cache) = self.receive_address_cache()? else {
-            let request_id = self.receive_address.next_request_id();
-            return self.open_fresh_receive_address(request_id, now);
-        };
-
-        if !self.receive_address_cache_is_available(&cache, now) {
-            let request_id = self.receive_address.next_request_id();
-            return self.open_fresh_receive_address(request_id, now);
-        }
-
-        let derivation_index = cache.derivation_index;
-        let request_id = self.receive_address.next_request_id();
-
-        if !matches!(
-            self.cached_receive_address_has_activity(derivation_index).await,
-            Ok(Some(true))
-        ) {
-            return self.open_cached_receive_address(cache, request_id, now).await;
-        }
-
-        if let Err(error) = self.sync_receive_address_now(derivation_index).await {
-            warn!("Failed to sync used receive address index={derivation_index}: {error}");
-            self.wallet.mark_receive_address_used(derivation_index)?;
-            self.notify_wallet_balance_and_transactions().await;
-
-            self.start_targeted_receive_address_sync(request_id, derivation_index);
-        }
-
-        self.open_fresh_receive_address(request_id, now)
-    }
-
-    fn receive_address_cache_is_available(
-        &self,
-        cache: &ReceiveAddressCache,
-        now_secs: u64,
-    ) -> bool {
-        let Some(expires_at_secs) = cache.first_shown_at_secs.checked_add(CACHE_WINDOW.as_secs())
-        else {
-            return false;
-        };
-
-        cache.address_type == self.wallet.metadata.address_type
-            && cache.first_shown_at_secs <= now_secs
-            && now_secs < expires_at_secs
-            && self.wallet.receive_address_is_unused(cache.derivation_index)
-    }
-
-    fn receive_address_cache(&self) -> Result<Option<ReceiveAddressCache>, Error> {
-        let cache = self.db.get_receive_address_cache().map_err_str(Error::ReceiveAddressError)?;
-
-        Ok(cache.filter(|cache| {
-            cache.wallet_id == self.wallet.id
-                && cache.network == self.wallet.network
-                && cache.address_type == self.wallet.metadata.address_type
-        }))
-    }
-
-    fn open_fresh_receive_address(
-        &mut self,
-        request_id: u64,
-        now: u64,
-    ) -> Result<ReceiveAddressState, Error> {
-        let state = self.new_receive_address_state(request_id, now, ReceiveAddressStatus::Fresh)?;
-
-        Ok(state)
-    }
-
-    async fn open_cached_receive_address(
-        &mut self,
-        cache: ReceiveAddressCache,
-        request_id: u64,
-        now: u64,
-    ) -> Result<ReceiveAddressState, Error> {
-        let cache = cache.with_visible_window_start(now);
-        let derivation_index = cache.derivation_index;
-        self.db.set_receive_address_cache(cache).map_err_str(Error::ReceiveAddressError)?;
-
-        let address = self.wallet.receive_address_at_index(derivation_index);
-        let state = ReceiveAddressState::cached(
-            request_id,
-            address,
-            ReceiveAddressStatus::CachedUnused,
-            now,
-        );
-        self.receive_address.set_visible(state.clone());
-        self.receive_address.set_refresh_state(ReceiveAddressRefreshState::Idle);
-        self.send_receive_address_state(state.clone());
-
-        self.start_receive_address_watcher(request_id, derivation_index);
-        self.schedule_receive_address_refresh(&state, now);
-        self.start_delayed_receive_address_activity_check(request_id, derivation_index);
-
-        Ok(state)
-    }
-
-    pub async fn create_new_receive_address_intent(&mut self) -> ActorResult<()> {
-        self.set_receive_address_loading(true);
-
-        match self.do_create_new_receive_address() {
-            Ok(_) => self.set_receive_address_loading(false),
-            Err(error) => {
-                self.set_receive_address_loading(false);
-                self.send(WalletManagerReconcileMessage::ReceiveAddressError(error.to_string()));
-            }
-        }
-
-        Produces::ok(())
-    }
-
-    fn do_create_new_receive_address(&mut self) -> Result<ReceiveAddressState, Error> {
-        let request_id = self.receive_address.next_request_id();
-        let now = current_epoch_secs();
-        let state = self.new_receive_address_state(request_id, now, ReceiveAddressStatus::Fresh)?;
-
-        Ok(state)
-    }
-
-    pub async fn refresh_expired_receive_address(&mut self, request_id: u64) -> ActorResult<()> {
-        let now = current_epoch_secs();
-        match self.receive_address.refresh_expired_decision(request_id, now) {
-            RefreshExpiredAddressDecision::Rotate => {}
-            RefreshExpiredAddressDecision::ReturnVisible(_)
-            | RefreshExpiredAddressDecision::MissingVisibleState => return Produces::ok(()),
-        }
-
-        self.stop_receive_address_refresh_timer();
-        self.set_receive_address_refresh_state(ReceiveAddressRefreshState::Refreshing);
-
-        match self.new_receive_address_state(request_id, now, ReceiveAddressStatus::Fresh) {
-            Ok(_) => Produces::ok(()),
-            Err(error) => {
-                if self.receive_address.visible_state().is_none() {
-                    return Err(Error::ReceiveAddressError(
-                        "receive request has no visible address".into(),
-                    )
-                    .into());
-                }
-
-                warn!("Failed to refresh receive address request_id={request_id}: {error}");
-                self.set_receive_address_refresh_state(ReceiveAddressRefreshState::Failed);
-                Produces::ok(())
-            }
-        }
-    }
-
-    pub async fn close_receive_address(&mut self, request_id: u64) {
-        if self.receive_address.close(request_id) {
-            self.stop_receive_address_watcher();
-            self.stop_receive_address_refresh_timer();
-            self.set_receive_address_loading(false);
-            self.send_receive_address_presentation();
-        }
-
-        self.send(WalletManagerReconcileMessage::ReceiveAddressClosed(request_id));
-    }
-
-    #[into_actor_result]
-    pub async fn check_node_connection(&mut self) {
-        let node = Database::global().global_config.selected_node();
-
-        let reconciler = self.reconciler.clone();
-        self.addr.send_fut(async move {
-            if let Err(error) = check_node_connection_inner(&node).await {
-                let _ = reconciler
-                    .send(WalletManagerReconcileMessage::NodeConnectionFailed(error).into());
-            }
-        });
     }
 
     pub async fn wallet_scan_and_notify(&mut self, force_scan: bool) -> ActorResult<()> {
@@ -1151,26 +271,6 @@ impl WalletActor {
         Produces::ok(())
     }
 
-    pub async fn get_height(&mut self, force: bool) -> ActorResult<usize> {
-        if let Some((last_height_fetched, block_height)) = self.last_height_fetched() {
-            let elapsed = elapsed_secs_since(last_height_fetched);
-            if !force && elapsed < 120 {
-                // if less than 25 seconds return the height, without updating
-                if elapsed < 25 {
-                    return Produces::ok(block_height);
-                }
-
-                // if more than a minute return height immediately, but update the height in the background
-                send!(self.addr.update_height());
-                return Produces::ok(block_height);
-            }
-        }
-
-        // update the height and return the new height
-        let block_height = self.update_height().await?.await?;
-        Produces::ok(block_height)
-    }
-
     pub async fn switch_mnemonic_to_new_address_type(
         &mut self,
         address_type: WalletAddressType,
@@ -1194,28 +294,6 @@ impl WalletActor {
         self.reset_scan_lifecycle_for_address_type_switch();
 
         Produces::ok(())
-    }
-
-    async fn update_height(&mut self) -> ActorResult<usize> {
-        debug!("actor update_height");
-        self.check_node_connection().await?;
-        let node_client = self.node_client().await?;
-        let block_height = node_client.get_height().await.map_err(|_| Error::GetHeightError)?;
-
-        self.save_last_height_fetched(block_height);
-        Produces::ok(block_height)
-    }
-
-    async fn update_block_id(&mut self) -> eyre::Result<BlockId> {
-        debug!("actor update_block_id");
-        if self.check_node_connection().await.is_err() {
-            return Err(eyre::eyre!("node connection failed"));
-        }
-
-        let node_client = self.node_client().await?;
-        let block_id = node_client.get_block_id().await.map_err(|_| Error::GetHeightError)?;
-        self.save_last_height_fetched(block_id.height as usize);
-        Ok(block_id)
     }
 
     #[into_actor_result]
@@ -1244,17 +322,7 @@ impl WalletActor {
     }
 
     pub async fn transaction_details(&mut self, tx_id: TxId) -> ActorResult<TransactionDetails> {
-        let tx = self
-            .wallet
-            .bdk
-            .get_tx(tx_id.0)
-            .ok_or(Error::TransactionDetailsError("transaction not found".to_string()))?;
-
-        let labels = self.db.labels.all_labels_for_txn(tx.tx_node.txid)?;
-        let details = TransactionDetails::try_new(&self.wallet.bdk, tx, labels.into())
-            .map_err_str(Error::TransactionDetailsError)?;
-
-        Produces::ok(details)
+        Produces::ok(self.transaction_details_for_tx_id(tx_id)?)
     }
 
     pub async fn current_wallet_unspent_outpoints_for_txn(
@@ -1275,49 +343,6 @@ impl WalletActor {
         Ok(state)
     }
 
-    pub async fn start_transaction_watcher(&mut self, tx_id: Txid) -> ActorResult<()> {
-        debug!("start_transaction_watcher for txn: {tx_id}");
-        if self.transaction_watchers.contains_key(&tx_id) {
-            warn!("transaction watcher already exists for txn: {tx_id}");
-            return Produces::ok(());
-        }
-
-        let network = self.wallet.network;
-        let node = Database::global().global_config.selected_node();
-        let options = NodeClientOptions { batch_size: 1 };
-        let client_builder = NodeClientBuilder { node, options };
-
-        let watcher = TransactionWatcher::new(self.addr.clone(), tx_id, client_builder, network);
-        let addr = spawn_actor(watcher);
-
-        self.transaction_watchers.insert(tx_id, addr);
-
-        Produces::ok(())
-    }
-
-    /// will remove the transaction watcher if it exists and
-    /// perform a sync scan and send the transactions to the frontend
-    pub async fn mark_transaction_found(&mut self, tx_id: Txid) -> ActorResult<()> {
-        info!("marking transaction found: {tx_id}");
-
-        // update the height
-        self.update_block_id().await?;
-
-        // update the height and perform sync scan which will update the transactions
-        self.perform_scan_for_single_tx_id(tx_id).await?;
-
-        // wait 30 seconds run the scan again and then remove the watcher
-        // sanity check to make sure the transaction was picked up by the wallet
-        // and no extra watchers were created in the meantime
-        self.addr.send_fut_with(|addr| async move {
-            tokio::time::sleep(Duration::from_secs(30)).await;
-            send!(addr.perform_delayed_scan_for_watched_tx_id(tx_id));
-            send!(addr.remove_watcher_for_txn(tx_id));
-        });
-
-        Produces::ok(())
-    }
-
     pub async fn shutdown(&mut self) {
         debug!("shutdown wallet actor");
         let scan_generation = self.advance_scan_generation();
@@ -1326,58 +351,18 @@ impl WalletActor {
             send!(scan_actor.shutdown(scan_generation));
         }
 
-        // cancel any in-flight payjoin negotiation and drop our strong actor handle
+        self.stop_receive_address_watcher();
+        self.stop_receive_address_refresh_timer();
         if let Some(actor) = self.payjoin_actor.take() {
             send!(actor.cancel_and_fallback());
         }
-        self.stop_receive_address_watcher();
-        self.stop_receive_address_refresh_timer();
         self.state = ActorState::Initial;
-        self.transaction_watchers = HashMap::default();
-        self.send_scan_idle_status();
-    }
-
-    async fn remove_watcher_for_txn(&mut self, tx_id: Txid) {
-        debug!("removing watcher for txn: {tx_id}");
-        self.transaction_watchers.remove(&tx_id);
-    }
-
-    pub async fn perform_scan_for_single_tx_id(&mut self, tx_id: Txid) -> ActorResult<()> {
-        let start = UNIX_EPOCH.elapsed().unwrap().as_secs();
-        let _ = self.update_height().await?.await;
-
-        let chain_tip = self.wallet.bdk.local_chain().tip();
-        let sync_request_builder = SyncRequest::builder().txids(vec![tx_id]).chain_tip(chain_tip);
-
-        let sync_request = sync_request_builder.build();
-
-        let node_client = self.node_client().await?.clone();
-        let graph = self.wallet.bdk.tx_graph().clone();
-
-        let now = UNIX_EPOCH.elapsed().unwrap().as_secs();
-        debug!("done scan for spk in {}s", now - start);
-        let generation = self.scan_generation;
-        self.addr.send_fut_with(|addr| async move {
-            let scan_result = node_client.sync(&graph, sync_request).await;
-
-            let now = UNIX_EPOCH.elapsed().unwrap().as_secs();
-            debug!("done single txn id sync scan in {}s", now - start);
-
-            // save updated txns and send to frontend
-            send!(addr.update_sync_state_and_send_transactions(scan_result, generation));
-        });
-
-        Produces::ok(())
-    }
-
-    async fn perform_delayed_scan_for_watched_tx_id(&mut self, tx_id: Txid) -> ActorResult<()> {
-        // shutdown clears watchers, so skip delayed scans for removed transactions
-        if !self.transaction_watchers.contains_key(&tx_id) {
-            debug!("skipping single tx scan, watcher already removed for {tx_id}");
-            return Produces::ok(());
+        for watcher in self.transaction_watchers.values() {
+            send!(watcher.stop_watching());
         }
 
-        self.perform_scan_for_single_tx_id(tx_id).await
+        self.transaction_watchers = HashMap::default();
+        self.send_scan_idle_status();
     }
 
     async fn perform_full_scan(&mut self) -> ActorResult<()> {
@@ -1392,128 +377,6 @@ impl WalletActor {
         send!(scan_actor.start_full_scan(self.scan_generation, ScanProgressStart::Immediate));
 
         Produces::ok(())
-    }
-
-    /// Get the max amount that can be sent for the given utxos
-    /// The total amount max is the total value of the UTXOs.
-    /// If the total amount is less than the UTXO total amount we will have a change output as well.
-    fn get_max_send_for_utxos(
-        &mut self,
-        total_amount: Amount,
-        address: &Address,
-        fee_rate: impl Into<BdkFeeRate>,
-        utxos: &[bitcoin::OutPoint],
-    ) -> Result<Amount, Error> {
-        let fee_rate = fee_rate.into();
-
-        // bdk manually selected UTXOs can bypass `unspendable`
-        self.reject_locked_outpoints(utxos)?;
-
-        let (utxo_total_amount, fee_estimate) = {
-            let mut utxo_total_amount = Amount::ZERO;
-            let mut total_fee_amount = Amount::ZERO;
-
-            let weighted_utxos =
-                self.get_weighted_utxos(utxos).map_err_str(Error::AddUtxosError)?;
-
-            for weighted_utxo in weighted_utxos {
-                let weight = TxIn::default()
-                    .segwit_weight()
-                    .checked_add(weighted_utxo.satisfaction_weight)
-                    .expect("`Weight` addition should not cause an integer overflow");
-
-                utxo_total_amount += weighted_utxo.utxo.txout().value;
-                total_fee_amount += fee_rate * weight;
-            }
-
-            (utxo_total_amount, total_fee_amount)
-        };
-
-        if total_amount > utxo_total_amount {
-            return Err(Error::InsufficientFunds(format!(
-                "custom amount {total_amount} is greater than the total amount available, total available: {utxo_total_amount}, fees: {fee_estimate}",
-            )));
-        }
-
-        let fee = {
-            let mut max_send_estimate =
-                utxo_total_amount.checked_sub(fee_estimate).ok_or_else(|| {
-                    Error::InsufficientFunds(format!(
-                        "no enough funds to cover the fees, total available: {total_amount}, fees: {fee_estimate}",
-                    ))
-                })?;
-
-            let mut fee_psbt = None;
-            while fee_psbt.is_none() {
-                if max_send_estimate < MIN_SEND_AMOUNT {
-                    return Err(Error::InsufficientFunds(format!(
-                        "no enough funds to cover the fees, total available: {total_amount}, fees: {fee_estimate}",
-                    )));
-                }
-
-                let mut tx_builder = self.wallet.bdk.build_tx();
-                tx_builder.add_utxos(utxos).map_err_str(Error::AddUtxosError)?;
-                tx_builder.manually_selected_only();
-                tx_builder.ordering(TxOrdering::Untouched);
-                tx_builder.add_recipient(address.script_pubkey(), max_send_estimate);
-                tx_builder.fee_rate(fee_rate);
-                let current_fee_psbt = tx_builder.finish();
-
-                match current_fee_psbt {
-                    Ok(psbt) => {
-                        fee_psbt = Some(psbt);
-                    }
-                    Err(CreateTxError::CoinSelection(result)) => {
-                        let difference = result.needed - result.available;
-                        max_send_estimate -= difference;
-                    }
-                    Err(err) => return Err(Error::BuildTxError(err.to_string())),
-                }
-            }
-
-            let fee_psbt = fee_psbt.expect("unwrapped in while");
-            self.wallet.unreserve_tx_change_addresses(&fee_psbt.unsigned_tx);
-            fee_psbt.fee().map_err_str(Error::BuildTxError)?
-        };
-
-        let max_send_amount = utxo_total_amount.checked_sub(fee).ok_or_else(|| {
-            Error::InsufficientFunds(format!(
-                "no enough funds to cover the fees, total available: {total_amount}, fees: {fee}",
-            ))
-        })?;
-
-        debug!("getting_max::send_amount: {max_send_amount:?}");
-        if total_amount < max_send_amount {
-            return Ok(total_amount);
-        }
-
-        Ok(max_send_amount)
-    }
-
-    fn get_weighted_utxos(
-        &self,
-        outpoints: &[bitcoin::OutPoint],
-    ) -> Result<Vec<WeightedUtxo>, AddUtxoError> {
-        outpoints
-            .iter()
-            .map(|outpoint| {
-                self.wallet.bdk.get_utxo(*outpoint).ok_or(AddUtxoError::UnknownUtxo(*outpoint)).map(
-                    |output| WeightedUtxo {
-                        satisfaction_weight: self
-                            .wallet
-                            .bdk
-                            .public_descriptor(output.keychain)
-                            .max_weight_to_satisfy()
-                            .unwrap(),
-                        utxo: Utxo::Local(output),
-                    },
-                )
-            })
-            .collect()
-    }
-
-    fn current_wallet_unspent_outpoints_for_txid(&self, txid: Txid) -> Vec<OutPoint> {
-        current_wallet_unspent_outpoints_for_txid(self.wallet.bdk.list_unspent(), txid)
     }
 
     fn lock_state_for_outpoints(
@@ -1764,37 +627,6 @@ impl WalletActor {
         Produces::ok(())
     }
 
-    async fn update_sync_state_and_send_transactions(
-        &mut self,
-        scan_result: Result<SyncResponse, NodeError>,
-        generation: WalletScanGeneration,
-    ) -> ActorResult<()> {
-        if generation != self.scan_generation {
-            debug!(
-                "dropping stale single tx scan result (gen {generation:?} != {:?})",
-                self.scan_generation
-            );
-            return Produces::ok(());
-        }
-
-        if scan_result.is_err() {
-            self.state = ActorState::FailedSyncScan;
-        }
-
-        let scan_result: SyncResponse = scan_result?;
-        self.wallet.bdk.apply_update(scan_result)?;
-        self.wallet.persist()?;
-        self.update_visible_receive_address_payment_status(None);
-
-        // get and send transactions
-        let transactions = self.transactions().await?.await?;
-        self.send(WalletManagerReconcileMessage::UpdatedTransactions(transactions));
-
-        self.state = ActorState::SyncScanComplete;
-
-        Produces::ok(())
-    }
-
     /// Mark the wallet as scanned
     /// Notify the frontend that the wallet scan is complete
     /// Ssend the wallet balance and transaction
@@ -1884,397 +716,11 @@ impl WalletActor {
     fn ensure_ledger_ready_for_spend(&self) -> Result<(), Error> {
         ledger_ready_for_spend(self.completed_initial_scan())
     }
-
-    fn last_height_fetched(&mut self) -> Option<(Duration, usize)> {
-        if let Some(last_height_fetched) = self.last_height_fetched {
-            return Some(last_height_fetched);
-        }
-
-        let metadata = Database::global()
-            .wallets()
-            .get(&self.wallet.id, self.wallet.network, self.wallet.metadata.wallet_mode)
-            .ok()??;
-
-        let BlockSizeLast { block_height, last_seen } = &metadata.internal.last_height_fetched?;
-
-        let last_height_fetched = Some((*last_seen, *(block_height) as usize));
-        self.last_height_fetched = last_height_fetched;
-
-        last_height_fetched
-    }
-
-    fn save_last_height_fetched(&mut self, block_height: usize) -> Option<()> {
-        debug!("actor save_last_height_fetched");
-        let now = UNIX_EPOCH.elapsed().unwrap_or_default();
-        self.last_height_fetched = Some((now, block_height));
-
-        let wallets = Database::global().wallets();
-        let mut metadata = wallets
-            .get(&self.wallet.id, self.wallet.network, self.wallet.metadata.wallet_mode)
-            .ok()??;
-
-        let last_height_fetched =
-            BlockSizeLast { block_height: block_height as u64, last_seen: now };
-
-        metadata.internal.last_height_fetched = Some(last_height_fetched);
-        wallets.update_internal_metadata(&metadata).ok();
-
-        Database::global()
-            .global_cache
-            .set_block_height(self.wallet.network, last_height_fetched)
-            .ok();
-
-        self.wallet.metadata = metadata.clone();
-
-        Some(())
-    }
-
-    fn new_receive_address_state(
-        &mut self,
-        request_id: u64,
-        now: u64,
-        status: ReceiveAddressStatus,
-    ) -> Result<ReceiveAddressState, Error> {
-        let address = self.wallet.get_next_address()?;
-        let cache = ReceiveAddressCache {
-            derivation_index: address.info.index,
-            first_shown_at_secs: now,
-            wallet_id: self.wallet.id.clone(),
-            network: self.wallet.network,
-            address_type: self.wallet.metadata.address_type,
-        };
-
-        self.db.set_receive_address_cache(cache).map_err_str(Error::ReceiveAddressError)?;
-
-        let state = ReceiveAddressState::cached(request_id, address, status, now);
-        self.receive_address.set_visible(state.clone());
-        self.receive_address.set_refresh_state(ReceiveAddressRefreshState::Idle);
-        self.send_receive_address_state(state.clone());
-
-        if status == ReceiveAddressStatus::Fresh {
-            self.start_receive_address_watcher(request_id, state.address.info.index);
-        } else {
-            self.stop_receive_address_watcher();
-        }
-
-        self.schedule_receive_address_refresh(&state, now);
-
-        Ok(state)
-    }
-
-    fn set_receive_address_loading(&self, is_loading: bool) {
-        self.send(WalletManagerReconcileMessage::ReceiveAddressLoadingChanged(is_loading));
-    }
-
-    fn set_receive_address_refresh_state(&mut self, refresh_state: ReceiveAddressRefreshState) {
-        self.receive_address.set_refresh_state(refresh_state);
-        self.send_receive_address_presentation();
-    }
-
-    fn send_receive_address_state(&self, state: ReceiveAddressState) {
-        self.send(WalletManagerReconcileMessage::ReceiveAddressUpdated(state));
-        self.send_receive_address_presentation();
-    }
-
-    fn send_receive_address_presentation(&self) {
-        let presentation: ReceiveAddressPresentation = self.receive_address.presentation();
-        self.send(WalletManagerReconcileMessage::ReceiveAddressPresentationUpdated(presentation));
-    }
-
-    fn schedule_receive_address_refresh(&mut self, state: &ReceiveAddressState, now_secs: u64) {
-        self.stop_receive_address_refresh_timer();
-
-        let Some(delay) = state.refresh_delay(now_secs) else {
-            return;
-        };
-
-        let request_id = state.request_id;
-        let addr = self.addr.clone();
-        self.receive_address_refresh_timer = Some(AbortableTask::spawn(async move {
-            tokio::time::sleep(delay).await;
-            send!(addr.refresh_expired_receive_address(request_id));
-        }));
-    }
-
-    fn stop_receive_address_refresh_timer(&mut self) {
-        self.receive_address_refresh_timer = None;
-    }
-
-    async fn cached_receive_address_has_activity(
-        &mut self,
-        derivation_index: u32,
-    ) -> Result<Option<bool>, Error> {
-        let node_client = self.node_client().await?.clone();
-        let address =
-            self.wallet.bdk.peek_address(KeychainKind::External, derivation_index).address;
-
-        match node_client
-            .check_address_for_txn(address)
-            .with_timeout(RECEIVE_ADDRESS_FRESHNESS_TIMEOUT)
-            .await
-        {
-            Ok(Ok(has_activity)) => Ok(Some(has_activity)),
-            Ok(Err(error)) => Err(Error::NodeConnectionFailed(error.to_string())),
-            Err(_) => Ok(None),
-        }
-    }
-
-    fn start_delayed_receive_address_activity_check(
-        &mut self,
-        request_id: u64,
-        derivation_index: u32,
-    ) {
-        let node = Database::global().global_config.selected_node();
-        let address =
-            self.wallet.bdk.peek_address(KeychainKind::External, derivation_index).address;
-        self.addr.send_fut_with(|addr| async move {
-            let result = match NodeClient::new(&node).await {
-                Ok(node_client) => node_client.check_address_for_txn(address).await,
-                Err(error) => Err(error),
-            };
-
-            send!(addr.handle_receive_address_activity_result(
-                request_id,
-                derivation_index,
-                result
-            ));
-        });
-    }
-
-    async fn handle_receive_address_activity_result(
-        &mut self,
-        request_id: u64,
-        derivation_index: u32,
-        result: Result<bool, NodeError>,
-    ) -> ActorResult<()> {
-        if !self.receive_address.is_current(request_id) {
-            return Produces::ok(());
-        }
-
-        if result.unwrap_or(false) {
-            self.start_targeted_receive_address_sync(request_id, derivation_index);
-        }
-
-        Produces::ok(())
-    }
-
-    pub async fn handle_receive_address_watch_activity(
-        &mut self,
-        request_id: u64,
-        derivation_index: u32,
-    ) -> ActorResult<()> {
-        if !self.receive_address.can_mark_payment_received(request_id, derivation_index) {
-            return Produces::ok(());
-        }
-
-        self.stop_receive_address_watcher();
-        self.db.delete_receive_address_cache().map_err_str(Error::ReceiveAddressError)?;
-
-        if self.wallet.receive_address_is_unused(derivation_index) {
-            self.wallet.mark_receive_address_used(derivation_index)?;
-        }
-
-        let Some(state) = self.receive_address.mark_payment_received(request_id, derivation_index)
-        else {
-            return Produces::ok(());
-        };
-
-        self.stop_receive_address_refresh_timer();
-        self.receive_address.set_refresh_state(ReceiveAddressRefreshState::Idle);
-        self.send_receive_address_state(state);
-        self.start_targeted_receive_address_sync(request_id, derivation_index);
-
-        Produces::ok(())
-    }
-
-    fn start_receive_address_watcher(&mut self, request_id: u64, derivation_index: u32) {
-        self.stop_receive_address_watcher();
-
-        let node = Database::global().global_config.selected_node();
-        let options = NodeClientOptions { batch_size: 1 };
-        let client_builder = NodeClientBuilder { node, options };
-
-        let address =
-            self.wallet.bdk.peek_address(KeychainKind::External, derivation_index).address;
-
-        let watcher = ReceiveAddressWatcher::new(
-            self.addr.clone(),
-            request_id,
-            derivation_index,
-            address,
-            client_builder,
-            CACHE_WINDOW,
-        );
-
-        self.receive_address_watcher = Some(spawn_actor(watcher));
-    }
-
-    fn stop_receive_address_watcher(&mut self) {
-        if let Some(watcher) = self.receive_address_watcher.take() {
-            send!(watcher.stop_watching());
-        }
-    }
-
-    pub async fn handle_receive_address_watcher_stopped(
-        &mut self,
-        request_id: u64,
-        derivation_index: u32,
-    ) -> ActorResult<()> {
-        if self.receive_address.is_current(request_id)
-            && self.receive_address.visible_state().is_some_and(|state| {
-                state.status == ReceiveAddressStatus::PaymentReceived
-                    || state.address.info.index == derivation_index
-            })
-        {
-            self.receive_address_watcher = None;
-        }
-
-        Produces::ok(())
-    }
-
-    fn start_targeted_receive_address_sync(&mut self, request_id: u64, derivation_index: u32) {
-        let (node, graph, sync_request) = self.receive_address_sync_inputs(derivation_index);
-        self.addr.send_fut_with(|addr| async move {
-            let result = match NodeClient::new(&node).await {
-                Ok(node_client) => node_client.sync(&graph, sync_request).await,
-                Err(error) => Err(error),
-            };
-
-            send!(addr.handle_receive_address_sync_result(request_id, derivation_index, result));
-        });
-    }
-
-    async fn sync_receive_address_now(&mut self, derivation_index: u32) -> Result<(), Error> {
-        let (node, graph, sync_request) = self.receive_address_sync_inputs(derivation_index);
-        let node_client = NodeClient::new(&node)
-            .await
-            .map_err_prefix("failed to create node client", Error::NodeConnectionFailed)?;
-
-        let sync_result =
-            node_client.sync(&graph, sync_request).await.map_err_str(Error::ReceiveAddressError)?;
-        self.wallet.bdk.apply_update(sync_result).map_err_str(Error::ReceiveAddressError)?;
-
-        if self.wallet.receive_address_is_unused(derivation_index) {
-            self.wallet.mark_receive_address_used(derivation_index)?;
-        } else {
-            self.wallet.persist()?;
-        }
-
-        self.notify_wallet_balance_and_transactions().await;
-
-        Ok(())
-    }
-
-    fn receive_address_sync_inputs(
-        &mut self,
-        derivation_index: u32,
-    ) -> (Node, TxGraph, SyncRequest<(KeychainKind, u32)>) {
-        let node = Database::global().global_config.selected_node();
-        let address =
-            self.wallet.bdk.peek_address(KeychainKind::External, derivation_index).address;
-        let script_pubkey = address.script_pubkey();
-        let chain_tip = self.wallet.bdk.local_chain().tip();
-
-        let sync_request = SyncRequest::builder()
-            .chain_tip(chain_tip)
-            .spks_with_indexes([((KeychainKind::External, derivation_index), script_pubkey)])
-            .build();
-
-        let graph = self.wallet.bdk.tx_graph().clone();
-
-        (node, graph, sync_request)
-    }
-
-    async fn handle_receive_address_sync_result(
-        &mut self,
-        request_id: u64,
-        derivation_index: u32,
-        result: Result<SyncResponse, NodeError>,
-    ) -> ActorResult<()> {
-        let sync_result = result
-            .inspect_err(|error| warn!("Address sync failed index={derivation_index}: {error}"))
-            .map_err_str(Error::ReceiveAddressError)?;
-
-        self.wallet.bdk.apply_update(sync_result)?;
-
-        if !self.receive_address.is_current(request_id) {
-            self.wallet.persist()?;
-            self.notify_wallet_balance_and_transactions().await;
-
-            return Produces::ok(());
-        }
-
-        if let Some(state) = self.receive_address.visible_state()
-            && state.address.info.index == derivation_index
-            && state.status != ReceiveAddressStatus::PaymentReceived
-            && self.wallet.receive_address_is_unused(state.address.info.index)
-        {
-            self.wallet.mark_receive_address_used(state.address.info.index)?;
-        }
-
-        self.wallet.persist()?;
-
-        self.update_visible_receive_address_payment_status(Some(derivation_index));
-        self.notify_wallet_balance_and_transactions().await;
-
-        Produces::ok(())
-    }
-
-    async fn notify_wallet_balance_and_transactions(&mut self) {
-        let balance = self.wallet.balance();
-        self.send(WalletManagerReconcileMessage::WalletBalanceChanged(balance.into()));
-
-        let transactions = self.do_transactions().await;
-        self.send(WalletManagerReconcileMessage::UpdatedTransactions(transactions));
-    }
-
-    fn update_visible_receive_address_payment_status(
-        &mut self,
-        derivation_index: Option<u32>,
-    ) -> bool {
-        let Some(state) = self.receive_address.visible_state() else {
-            return false;
-        };
-
-        if state.status == ReceiveAddressStatus::PaymentReceived
-            || derivation_index.is_some_and(|index| state.address.info.index != index)
-            || self.wallet.receive_address_is_unused(state.address.info.index)
-        {
-            return false;
-        }
-
-        let state = state.payment_received();
-        self.receive_address.set_visible(state.clone());
-        self.receive_address.set_refresh_state(ReceiveAddressRefreshState::Idle);
-        self.stop_receive_address_watcher();
-        self.stop_receive_address_refresh_timer();
-        self.send_receive_address_state(state);
-
-        true
-    }
-
-    async fn node_client(&mut self) -> Result<&NodeClient, Error> {
-        let node_client = self.node_client.as_ref();
-        if node_client.is_none() {
-            let node = Database::global().global_config.selected_node();
-            let node_client = NodeClient::new(&node)
-                .await
-                .map_err_prefix("failed to create node client", Error::NodeConnectionFailed)?;
-
-            self.node_client = Some(node_client);
-        }
-
-        Ok(self.node_client.as_ref().expect("just checked"))
-    }
 }
 
 fn elapsed_secs_since(earlier: Duration) -> u64 {
     let now = UNIX_EPOCH.elapsed().unwrap_or(earlier);
     now.saturating_sub(earlier).as_secs()
-}
-
-fn current_epoch_secs() -> u64 {
-    UNIX_EPOCH.elapsed().unwrap_or_default().as_secs()
 }
 
 fn progressive_scan_update_response(
@@ -2471,11 +917,6 @@ impl WalletActor {
             self.scan_actor = None;
         }
     }
-
-    fn spawn_payjoin_actor(&mut self, sender: PayjoinSender, fallback_tx: BdkTransaction) {
-        self.payjoin_actor =
-            Some(spawn_actor(PayjoinActor::new(self.addr.clone(), sender, fallback_tx)));
-    }
 }
 
 impl Drop for WalletActor {
@@ -2564,37 +1005,11 @@ fn exclude_locked_outpoints<Cs>(
     tx_builder.unspendable(locked_outpoints);
 }
 
-async fn check_node_connection_inner(node: &Node) -> Result<(), String> {
-    // Create a fresh client with its own TCP connection for this background check.
-    // We cannot reuse the actor's cached client because:
-    // 1. This runs in a background task (spawned via send_fut)
-    // 2. The actor continues processing messages with its own client
-    // 3. The underlying rust-electrum-client is NOT designed for concurrent access
-    //    (it uses a "reader thread" pattern with try_lock that fails on concurrent use)
-    // Creating a fresh connection ensures no shared state or concurrent access.
-    //
-    // TODO: We could optimize this to reuse the cached client when using esplora,
-    // since esplora uses HTTP and doesn't have the concurrent access limitations
-    // that electrum's persistent TCP connection has.
-    let node_client = NodeClient::new(node)
-        .await
-        .map_err(|_| "unable to create a connection to the node".to_string())?;
-
-    node_client
-        .check_url()
-        .with_timeout(Duration::from_secs(5))
-        .await
-        .map_err(|_| "unable to connect to node, timeout".to_string())?
-        .map_err(|err| err.to_string())?;
-
-    Ok(())
-}
-
 #[cfg(test)]
 mod tests {
     use bdk_wallet::{
         KeychainKind, LocalOutput,
-        chain::{ChainPosition, ConfirmationBlockTime},
+        chain::{BlockId, ChainPosition, ConfirmationBlockTime},
         test_utils::{
             ReceiveTo, get_funded_wallet_wpkh, insert_checkpoint, receive_output,
             receive_output_in_latest_block, receive_output_to_address,
@@ -2615,7 +1030,7 @@ mod tests {
     use crate::wallet::metadata::WalletMetadata;
 
     use super::{
-        ActorState, BlockId, EMPTY_WALLET_SCAN_PROGRESS_DELAY, FullScanType, InitialScanRoute,
+        ActorState, EMPTY_WALLET_SCAN_PROGRESS_DELAY, FullScanType, InitialScanRoute,
         RETURNING_WALLET_SCAN_PROGRESS_DELAY, ScanProgressStart,
         full_scan_updates_initial_metadata, initial_scan_route, ledger_ready_for_spend,
         metadata_with_full_scan_performed, progressive_scan_update_response,
